Agrego el inicio de la comunicacion cliente-servidor. Por ahora lo unico que logre hacer es que se envien mensajes de conectado y todo ok, pero nada mas. Estuve tratando de hacer que el ServerProxy del cliente envie un mensaje de a donde tiene que moverse y el sv lea y diga, izq o derecha, pero no funciona :(. Ahora hay que seguir revisando. Comente las lineas del lint porque es muy rompebolas y no me deja compilar nada (espero que los warnings del lint no sean el motivo por el cual no anda la comunicacion)

+++ client/MainWindow.h
@@ -62,7 +62,7 @@ class MainWindow: public Gtk::Window {
         add(map_name);
       }
 
-      Gtk::TreeModelColumn<int> id;
+      Gtk::TreeModelColumn<size_t> id;
       Gtk::TreeModelColumn<Glib::ustring> map_name;
   };
